Fix hangup (race condition)
Our test application sometimes ends up in a kind of hangup.
This happens quite often with clang-14 release builds, seldomly with gcc-10
release builds, not with debug builds.

The hangup is a kind of busy-waiting state where two threads from a static
thread pool are continually trying to wake each other, but not really
succeeding because both are already awake. A third thread, that could
allow the application to proceed, doesn't get the chance to continue.

The cause seems to be that m_sleepingThreadCount is not reflecting the
number of sleeping threads after a notify_intent_to_sleep() call where
m_isSleeping was already set. This probably can happen due to a spurious
wake-up (see comment above the try_clear_intent_to_sleep() call).

The proposed solution ensures that m_sleepingThreadCount is not increased
in such a situation.

https://github.com/mghro/cradle/blob/reproduce-hangup/README.md
describes a reproduction scenario.